Add embed button to video card when embed link is present
[#184757456]
