+++ src/components/sparks/themePage/theme/SparksVideoCardButtons.tsx
@@ -1,6 +1,5 @@
 import React from 'react';
 import { Video } from 'boclips-api-client/dist/sub-clients/videos/model/Video';
-import { EmbedButton } from 'src/components/embedButton/EmbedButton';
 import { VideoCardButtons } from 'src/components/videoCard/buttons/VideoCardButtons';
 import { DownloadTranscriptButton } from 'src/components/downloadTranscriptButton/DownloadTranscriptButton';
 
@@ -9,9 +8,6 @@ interface Props {
 }
 
 export const SparksVideoCardButtons = ({ video }: Props) => {
-  const showEmbedButton = video.links.createEmbedCode;
-  const primaryButton = showEmbedButton && <EmbedButton video={video} />;
-
   const secondaryButton = video.links.transcript && (
     <DownloadTranscriptButton video={video} />
   );
@@ -19,7 +15,6 @@ export const SparksVideoCardButtons = ({ video }: Props) => {
   return (
     <VideoCardButtons
       additionalSecondaryButtons={secondaryButton}
-      primaryButton={primaryButton}
       video={video}
       iconOnly
     />
